consolidate routes, streamline blueprint imports
