+++ bapa/__init__.py
@@ -13,12 +13,9 @@ db = SQLAlchemy(app)
 mail = Mail(app)
 
 # blueprints
-from bapa.modules import home, membership, password, officers, admin
-app.register_blueprint(home.home_bp)
-app.register_blueprint(membership.memb_bp, url_prefix='/membership')
-app.register_blueprint(password.pass_bp, url_prefix='/password')
-app.register_blueprint(officers.offic_bp, url_prefix='/officers')
-app.register_blueprint(admin.admin_bp, url_prefix='/admin')
+from bapa.modules import blueprints
+for prefix, bp in blueprints.items():
+    app.register_blueprint(bp, url_prefix=prefix)
 
 # Template Globals
 app.jinja_env.globals['year'] = date.today().year
