+++ bapa/modules/__init__.py
@@ -1,0 +1,9 @@
+from . import admin, home, membership, officers
+
+blueprints = {
+    #url_prefix: Blueprint
+    '/admin' : admin.bp,
+    '': home.bp,
+    '/membership': membership.bp,
+    '/officers': officers.bp
+}
