+++ bapa/modules/admin/__init__.py
@@ -1,1 +1,1 @@
-from .routes import admin_bp
+from .routes import bp
