+++ bapa/modules/home/__init__.py
@@ -1,1 +1,1 @@
-from .routes import home_bp
+from .routes import bp
