+++ bapa/modules/home/routes.py
@@ -1,20 +1,19 @@
 from . import controllers
+from bapa.utils import timestamp, is_too_old
 from flask import render_template, redirect, url_for, flash
 from flask import session, request
 from flask import Blueprint
+import markdown2, os
 
-import markdown2
-import os
-
-home_bp = Blueprint('home', __name__, template_folder='templates')
+bp = Blueprint('home', __name__, template_folder='templates')
 
 
-@home_bp.route('/')
+@bp.route('/')
 def index():
     return render_template('home.html', session=session)
 
 
-@home_bp.route('/register', methods=['GET', 'POST'])
+@bp.route('/register', methods=['GET', 'POST'])
 def register():
     """Register the user."""
     if session.get('user'):
@@ -35,8 +34,7 @@ def register():
     return render_template('register.html', error=error)
 
 
-
-@home_bp.route('/login', methods=['GET', 'POST'])
+@bp.route('/login', methods=['GET', 'POST'])
 def login():
     """User login"""
     if session.get('user'):
@@ -53,8 +51,7 @@ def login():
     return render_template('login.html', error=error, session=session)
 
 
-
-@home_bp.route('/logout')
+@bp.route('/logout')
 def logout(msg='You were logged out'):
     """Logs the user out."""
     flash(msg)
@@ -62,7 +59,82 @@ def logout(msg='You were logged out'):
     return redirect(url_for('home.index'))
 
 
-@home_bp.route('/page/<name>')
+
+##################
+# Password Reset #
+##################
+
+@bp.route('/password/reset/request', methods=['GET', 'POST'])
+def reset_request():
+    if session.get('user'):
+        return redirect(url_for('home.index'))
+    error = None
+    if request.method == 'POST':
+        error = controllers.reset_password_request(
+            request.form['ushpa'],
+            request.form['email'],
+            url_for('home.reset_auth')
+        )
+        if not error:
+            flash('Email sent')
+            return redirect(url_for('home.index'))
+    return render_template('reset_req.html', error=error)
+
+
+@bp.route('/password/reset/auth/')
+@bp.route('/password/reset/auth/<secret>', methods=['GET', 'POST'])
+def reset_auth(secret=None):
+    if session.get('user') or not secret:
+        return redirect(url_for('home.index'))
+    if request.method == 'POST':
+        user = controllers.reset_password_auth(
+            request.form['ushpa'],
+            request.form['email'],
+            secret
+        )
+        if not user:
+            flash('Password Reset Failed')
+            return redirect(url_for('home.login'))
+        session['user'] = {}
+        session['user']['id'] = user.id
+        session['authed'] = timestamp(object=True)
+        return redirect(url_for('home.reset'))
+    return render_template('reset_auth.html', secret=secret)
+
+
+@bp.route('/password/auth', methods=['GET', 'POST'])
+def simple_auth():
+    if not session.get('user'):
+        return redirect(url_for('home.login'))
+    error = None
+    if request.method == 'POST':
+        if controllers.auth(session['user']['ushpa'], request.form['password']):
+            session['authed'] = timestamp(object=True)
+            return redirect(url_for('home.reset'))
+        error = 'Enter your current password'
+    return render_template('auth.html', error=error)
+
+
+@bp.route('/password/reset', methods=['GET', 'POST'])
+def reset():
+    if not session.get('user'):
+        return redirect(url_for('home.login'))
+    if is_too_old(session.get('authed')):
+        return redirect(url_for('home.simple_auth'))
+    error = None
+    if request.method == 'POST':
+        error = controllers.reset_password(
+            session['user']['id'],
+            request.form['password'],
+            request.form['password2']
+        )
+        if not error:
+            flash('Your password has been reset')
+            return redirect(url_for('home.logout'))
+    return render_template('reset.html', error=error)
+
+
+@bp.route('/page/<name>')
 def page(name=None):
     path = os.path.join(os.getcwd(), 'bapa', 'content', name + '.md')
 
@@ -79,7 +151,7 @@ def page(name=None):
     return render_template('pages/page.html', title=title, content=content)
 
 
-@home_bp.route('/news', methods=['GET'])
+@bp.route('/news', methods=['GET'])
 def news():
     page = request.args.get('page')
     n=3 #per page
