+++ bapa/modules/home/templates/auth.html
@@ -1,5 +1,5 @@
 {% extends "layout.html" %}
-{% block title %}Reset Password{% endblock %}
+{% block title %}Authenticate{% endblock %}
 {% block body %}
 
     <div class="container">
@@ -14,23 +14,18 @@
 
                 <div class="panel panel-default">
                     <div class="panel-heading">
-                        <h3 class="panel-title">Reset Password</h3>
+                        <h3 class="panel-title">Authenticate</h3>
                     </div>
                     <div class="panel-body">
-                        <form action="{{ url_for('password.reset_request') }}" method="POST">
+                        <form action="{{ url_for('home.simple_auth') }}" method="POST">
 
                             <div class="form-group">
-                                <label for="ushpa">USHPA Pilot Number</label>
-                                <input type="text" name="ushpa" id="ushpa" class="form-control">
+                                <label for="password">Enter Password: </label>
+                                <input type="password" name="password" id="password" class="form-control">
                             </div>
 
                             <div class="form-group">
-                                <label for="email">Email</label>
-                                <input type="email" name="email" id="email" class="form-control">
-                            </div>
-
-                            <div class="form-group">
-                                <input type="submit" value="Send Reset Email" class="btn btn-primary pull-right">
+                                <input type="submit" value="Authenticate" class="btn btn-primary pull-right">
                             </div>
                         </form>
                     </div>
