+++ bapa/modules/home/templates/login.html
@@ -30,7 +30,7 @@
 							</div>
 
 							<div class="form-group">
-								<a href="{{ url_for('password.reset_request') }}" class="btn btn-default">Forgot Password?</a>
+								<a href="{{ url_for('home.reset_request') }}" class="btn btn-default">Forgot Password?</a>
 								<input type="submit" value="Sign In" class="btn btn-primary pull-right">
 							</div>
 						</form>
