+++ bapa/modules/home/templates/reset.html
@@ -14,19 +14,19 @@
 
                 <div class="panel panel-default">
                     <div class="panel-heading">
-                        <h3 class="panel-title">Reenter Credentials</h3>
+                        <h3 class="panel-title">Reset Password</h3>
                     </div>
                     <div class="panel-body">
-                        <form action="{{ url_for('password.reset_auth', secret=secret) }}" method="POST">
+                        <form action="{{ url_for('home.reset') }}" method="POST">
 
                             <div class="form-group">
-                                <label for="ushpa">USHPA Pilot Number: </label>
-                                <input type="text" name="ushpa" id="ushpa" class="form-control">
+                                <label for="password">New Password: </label>
+                                <input type="password" name="password" id="password" class="form-control">
                             </div>
 
                             <div class="form-group">
-                                <label for="email">Email: </label>
-                                <input type="email" name="email" id="email" class="form-control">
+                                <label for="password2">New Password (verify): </label>
+                                <input type="password" name="password2" id="password2" class="form-control">
                             </div>
 
                             <div class="form-group">
