+++ bapa/modules/home/templates/reset_req.html
@@ -17,20 +17,20 @@
                         <h3 class="panel-title">Reset Password</h3>
                     </div>
                     <div class="panel-body">
-                        <form action="{{ url_for('password.reset') }}" method="POST">
+                        <form action="{{ url_for('home.reset_request') }}" method="POST">
 
                             <div class="form-group">
-                                <label for="password">New Password: </label>
-                                <input type="password" name="password" id="password" class="form-control">
+                                <label for="ushpa">USHPA Pilot Number</label>
+                                <input type="text" name="ushpa" id="ushpa" class="form-control">
                             </div>
 
                             <div class="form-group">
-                                <label for="password2">New Password (verify): </label>
-                                <input type="password" name="password2" id="password2" class="form-control">
+                                <label for="email">Email</label>
+                                <input type="email" name="email" id="email" class="form-control">
                             </div>
 
                             <div class="form-group">
-                                <input type="submit" value="Confirm" class="btn btn-primary pull-right">
+                                <input type="submit" value="Send Reset Email" class="btn btn-primary pull-right">
                             </div>
                         </form>
                     </div>
