+++ bapa/modules/officers/__init__.py
@@ -1,1 +1,1 @@
-from .routes import offic_bp
+from .routes import bp
