--- bapa/modules/password/__init__.py
@@ -1,1 +1,0 @@
-from .routes import pass_bp
